Extracted reportgenerator tests to class
