+++ src/RefRestrictTest/ConfigParserTest.cs
@@ -84,7 +84,7 @@ namespace RefRestrictTest
         }
     }
     [TestClass]
-    public class ProjectFileParserTest
+    public partial class ProjectFileParserTest
     {
         [TestMethod]
 
@@ -99,43 +99,6 @@ namespace RefRestrictTest
 
             //Assert
             Assert.IsTrue(projectInfo.NugetRefs.Count > 0);
-        }
-        [TestClass]
-        public class EndToEndTests
-        {
-            [TestMethod]
-            public void CreateEmptyReportWhenRulesAdheredToForPackageReference()
-            {
-                // Arrange
-                var configFile = @"TestConfigFiles\TestConfig_IncludePackageReference.xml";
-                var project = @"TestConfigFiles\RR.WebServer.csproj";
-                var projectInfo = ProjectFileParser.GetProjectInfo(project);
-                var ruleSet = ConfigParser.GetRuleSetForProject(configFile, projectInfo.Name);
-
-                //Act
-                var results = RefAnaylser.GenerateReport(ruleSet, projectInfo);
-
-                //Assert
-                Assert.IsTrue(results.Entries.Count == 0);
-
-            }
-
-            [TestMethod]
-            public void CreateNonEmptyReportWhenRulesAdheredToForPackageReference()
-            {
-                // Arrange
-                var configFile = @"TestConfigFiles\TestConfig_ExcludePackageReference.xml";
-                var project = @"TestConfigFiles\RR.WebServer.csproj";
-                var projectInfo = ProjectFileParser.GetProjectInfo(project);
-                var ruleSet = ConfigParser.GetRuleSetForProject(configFile, projectInfo.Name);
-
-                //Act
-                var results = RefAnaylser.GenerateReport(ruleSet, projectInfo);
-
-                //Assert
-                Assert.IsTrue(results.Entries.Count != 0);
-
-            }
         }
     }
 }
