+++ src/RefRestrictTest/RefRestrictTest.csproj
@@ -52,6 +52,7 @@
   <ItemGroup>
     <Compile Include="ConfigParserTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ReportGeneratorTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\RefRestrict\RefRestrict.csproj">
@@ -61,6 +62,9 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="TestConfigFiles\RR.WebServer.csproj">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
+    <Content Include="TestConfigFiles\TestConfig_ExcludePackageReference.xml">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
     <Content Include="TestConfigFiles\TestConfig_IncludePackageReference.xml">
